Fix queue index, add next/prev to queue, add api context,
