+++ app/src-tauri/tauri.conf.json
@@ -67,7 +67,7 @@
         "title": "Music Caster",
         "label": "main",
         "minWidth": 1000,
-        "minHeight": 700,
+        "minHeight": 710,
         "resizable": true,
         "fullscreen": false,
         "visible": false,
