+++ app/src/views/Queue.tsx
@@ -1,9 +1,10 @@
 import { Box, Flex, Paper, ScrollArea, Skeleton, Stack, Text } from '@mantine/core';
 import { useContext, useEffect, useRef } from 'react';
-import { PlayerStateContext } from '../common/contexts';
+import { MusicCasterAPIContext, PlayerStateContext } from '../common/contexts';
 
 export default function Queue() {
 	const playerState = useContext(PlayerStateContext);
+	const api = useContext(MusicCasterAPIContext)!;
 	const scrolledRef = useRef(false);
 	const currentTrackRef = useRef<HTMLDivElement>(null);
 
@@ -30,6 +31,15 @@ export default function Queue() {
 		);
 	}
 
+	const onTrackClick = (index: number) => {
+		if (index < 0) {
+			api.prev(-index);
+		}
+		else if (index > 0) {
+			api.next(index);
+		}
+	}
+
 	return (
 		<Box style={{ height: 'calc(100vh - 150px)' }}>
 			<Paper shadow="sm" p="md" style={{ height: '100%', overflow: 'hidden' }}>
@@ -48,10 +58,11 @@ export default function Queue() {
 										cursor: 'pointer',
 										backgroundColor: index === playerState.queue_position ? 'var(--mantine-color-blue-light)' : undefined
 									}}
+									onClick={() => onTrackClick(index - playerState.queue_position)}
 								>
 									<Flex gap="md" align="center">
 										<Text size="sm" c="dimmed" style={{ minWidth: '2em', textAlign: 'right' }}>
-											{index - playerState.queue_position + 1}
+											{index - playerState.queue_position}
 										</Text>
 										<Text size="sm" fw={500}>{track}</Text>
 									</Flex>
